+++ internal/services/messenger/internal/handler/incoming/push_test.go
@@ -159,7 +159,7 @@ func newTsHLSClient(pGetPubKey, pWithOK bool) *tsHLSClient {
 	}
 }
 
-func (p *tsHLSClient) GetIndex(context.Context) (string, error) { return "", nil }
+func (p *tsHLSClient) GetIndex(context.Context) error { return nil }
 func (p *tsHLSClient) GetSettings(context.Context) (hls_config.IConfigSettings, error) {
 	if !p.fWithOK {
 		return nil, errors.New("some error") // nolint: err113