+++ internal/services/remoter/pkg/settings/types.go
@@ -1,6 +1,0 @@
-package settings
-
-type SCommandExecRequest struct {
-	FPassword string   `json:"password"`
-	FCommand  []string `json:"command"`
-}