+++ pkg/api/services/filesharer/client/requester.go
@@ -2,8 +2,10 @@ package client
 
 import (
 	"context"
+	"crypto/sha512"
 	"errors"
 	"fmt"
+	"hash"
 	"io"
 	"net/http"
 	"net/url"
@@ -32,6 +34,25 @@ type sRequester struct {
 	fClient *http.Client
 }
 
+type sWriterWithHash struct {
+	fMain io.Writer
+	fHash hash.Hash
+}
+
+func newWriterWithHash(pW io.Writer) *sWriterWithHash {
+	return &sWriterWithHash{
+		fMain: pW,
+		fHash: sha512.New384(),
+	}
+}
+
+func (p *sWriterWithHash) Write(b []byte) (n int, err error) {
+	if _, err := p.fHash.Write(b); err != nil {
+		return 0, err
+	}
+	return p.fMain.Write(b)
+}
+
 func NewRequester(pHost string, pClient *http.Client) IRequester {
 	return &sRequester{
 		fHost:   pHost,
@@ -39,7 +60,7 @@ func NewRequester(pHost string, pClient *http.Client) IRequester {
 	}
 }
 
-func (p *sRequester) GetIndex(pCtx context.Context) (string, error) {
+func (p *sRequester) GetIndex(pCtx context.Context) error {
 	res, err := api.Request(
 		pCtx,
 		p.fClient,
@@ -48,15 +69,12 @@ func (p *sRequester) GetIndex(pCtx context.Context) (string, error) {
 		nil,
 	)
 	if err != nil {
-		return "", errors.Join(ErrBadRequest, err)
+		return errors.Join(ErrBadRequest, err)
 	}
-
-	result := string(res)
-	if result != hls_settings.CAppFullName {
-		return "", ErrInvalidTitle
+	if string(res) != hls_settings.CAppFullName {
+		return ErrInvalidTitle
 	}
-
-	return result, nil
+	return nil
 }
 
 func (p *sRequester) GetRemoteList(pCtx context.Context, pAliasName string, pPage uint64, pPersonal bool) (fileinfo.IFileInfoList, error) {
@@ -77,9 +95,10 @@ func (p *sRequester) GetRemoteList(pCtx context.Context, pAliasName string, pPag
 	return infos, nil
 }
 
-func (p *sRequester) GetRemoteFile(pW io.Writer, pCtx context.Context, pAliasName string, pFileName string, pPersonal bool) (bool, string, error) {
+func (p *sRequester) GetRemoteFile(pW io.Writer, pCtx context.Context, pAliasName string, pFileName string, pPersonal bool) (bool, error) {
+	writer := newWriterWithHash(pW)
 	headers, err := api.RequestWithWriter(
-		pW,
+		writer,
 		pCtx,
 		p.fClient,
 		http.MethodGet,
@@ -87,10 +106,24 @@ func (p *sRequester) GetRemoteFile(pW io.Writer, pCtx context.Context, pAliasNam
 		nil,
 	)
 	if err != nil {
-		return false, "", errors.Join(ErrBadRequest, err)
+		return false, errors.Join(ErrBadRequest, err)
 	}
-	alreadyDownload := headers.Get(hls_settings.CHeaderInProcess) == hls_settings.CHeaderProcessModeY
-	return alreadyDownload, headers.Get(hls_settings.CHeaderFileHash), nil
+	inProcess := headers.Get(hls_settings.CHeaderInProcess) == hls_settings.CHeaderProcessModeY
+	return inProcess, nil
+}
+
+func (p *sRequester) DelRemoteFile(pCtx context.Context, pAliasName string, pFileName string, pPersonal bool) error {
+	_, err := api.Request(
+		pCtx,
+		p.fClient,
+		http.MethodDelete,
+		fmt.Sprintf(cHandleRemoteFileTemplate, p.fHost, url.QueryEscape(pAliasName), url.QueryEscape(pFileName), pPersonal),
+		nil,
+	)
+	if err != nil {
+		return errors.Join(ErrBadRequest, err)
+	}
+	return nil
 }
 
 func (p *sRequester) GetRemoteFileInfo(pCtx context.Context, pAliasName string, pFileName string, pPersonal bool) (fileinfo.IFileInfo, error) {