+++ CHANGELOG.md
@@ -2,9 +2,28 @@
 
 <!-- ... -->
 
-## v1.10.2~
+## v1.10.3~
 
 *??? ??, ????*
+
+<!-- ... -->
+
+## v1.10.2
+
+*February 11, 2026*
+
+### IMPROVEMENTS
+
+- `cmd/hls/hls-filesharer`: add API function DelRemoteFile
+
+### CHANGES
+
+- `cmd/hls/hls-remoter`: deleted
+- `cmd/hls/hls-filesharer`: change API function GetRemoteFile
+- `cmd/hls/hls-filesharer`: now process downloading checker start before getFileInfo
+- `cmd/hls/hls-filesharer`: The file names in private are now different when downloading a file with the same name from the personal and sharing directories
+- `pkg/api/*/client`: change GetIndex functions
+- `cmd/hlc|cmd/hlk`: add hls-filesharer to init config
 
 <!-- ... -->